Fix warning under MSVC.
FossilOrigin-Name: 0c32c4bbdd74297767dcf4ec4295f9cc72875af0

+++ src/wal.c
@@ -1458,7 +1458,7 @@ static int walIteratorInit(Wal *pWal, WalIterator **pp){
       if( (i+1)==nSegment ){
         nEntry = (int)(iLast - iZero);
       }else{
-        nEntry = (u32*)aHash - (u32*)aPgno;
+        nEntry = (int)((u32*)aHash - (u32*)aPgno);
       }
       aIndex = &((ht_slot *)&p->aSegment[p->nSegment])[iZero];
       iZero++;
